Removed inconsistency in setting baud rate

+++ src/jtermios/macosx/JTermiosImpl.java
@@ -185,12 +185,12 @@ public class JTermiosImpl implements jtermios.JTermios.JTermiosInterface {
 		return m_Clib.tcdrain(fd);
 	}
 
-	public long cfgetispeed(Termios termios) {
-		return m_Clib.cfgetispeed(new MacOSX_C_lib.Termios(termios)).longValue();
+	public int cfgetispeed(Termios termios) {
+		return m_Clib.cfgetispeed(new MacOSX_C_lib.Termios(termios)).intValue();
 	}
 
-	public long cfgetospeed(Termios termios) {
-		return m_Clib.cfgetospeed(new MacOSX_C_lib.Termios(termios)).longValue();
+	public int cfgetospeed(Termios termios) {
+		return m_Clib.cfgetospeed(new MacOSX_C_lib.Termios(termios)).intValue();
 	}
 
 	public int cfsetispeed(Termios termios, int speed) {
